Attach a model to a pipeline

The model is the final stage of the pipeline. It will be fitted on the transformed training series when milt_pipeline_fit() is called.

Usage

milt_pipe_model(pipeline, model, ...)

Arguments

pipeline

A MiltPipeline.

model

Either a MiltModel object (from milt_model()) or a character model name (e.g. "auto_arima").

...

Hyperparameters forwarded to milt_model() when model is a character string.

Value

The updated MiltPipeline, invisibly.

See also

Examples

# \donttest{
pipe <- milt_pipeline() |>
  milt_pipe_step_lag(lags = 1:6) |>
  milt_pipe_model("naive")
# }